requirements.txt
numpy>=1.19.0
pandas>=1.1.0
matplotlib>=3.3.0
seaborn>=0.11.0
scipy>=1.5.0
networkx>=2.5.0
tensorflow==2.17.0
scikit-learn>=0.24.0
pillow>=8.0.0
customtkinter>=4.6.0
tqdm>=4.50.0